fix: commands pages take up full width available
Fixes #351. The `<section>` for the commands pages already has the
`w-full` class but the containing div did not, meaning the central
content of each page would not take up the full width available. This
commit fixes this issue by adding the `w-full` class to the containing
div.

These changes were tested to function on Firefox on Linux and Chrome on
macOS, with additional verification that the pages looked the same on
Safari on iOS and Chrome on Android.