Update installing-redisgears.md
Adding directions on how to use the newly installed redisgears version to upgrade a database. 
It was a little jarring not having this available on this page. It left me hanging when I needed to upgrade a database. It took me a good five minutes to fine the information on how to apply the upgrade to a database.