Update installing-redisgears.md (#1697)
* Update installing-redisgears.md

Adding directions on how to use the newly installed redisgears version to upgrade a database. 
It was a little jarring not having this available on this page. It left me hanging when I needed to upgrade a database. It took me a good five minutes to fine the information on how to apply the upgrade to a database.

* Copy edits for upgrading RedisGears

* Noted that the example commands also upgrade the Redis DB version

---------

Co-authored-by: Rachel Elledge <86307637+rrelledge@users.noreply.github.com>